Fixed up Meta endpoint again to use the database structure
 - and cleaned up the code structure a good bit too.

+++ src/main/java/org/cru/godtools/api/meta/MetaPackage.java
@@ -1,6 +1,7 @@
 package org.cru.godtools.api.meta;
 
 import javax.xml.bind.annotation.XmlAttribute;
+import java.math.BigDecimal;
 
 /**
  * Created by ryancarlson on 3/27/14.
@@ -8,14 +9,14 @@ import javax.xml.bind.annotation.XmlAttribute;
 public class MetaPackage  implements java.io.Serializable
 {
     String code;
-    Integer version;
+	BigDecimal version;
     String name;
 
     public MetaPackage()
     {
     }
 
-    public MetaPackage(String packageName, String packageCode, Integer versionNumber)
+    public MetaPackage(String packageName, String packageCode, BigDecimal versionNumber)
     {
         setName(packageName);
         setCode(packageCode);
@@ -34,12 +35,12 @@ public class MetaPackage  implements java.io.Serializable
     }
 
     @XmlAttribute
-    public Integer getVersion()
+    public BigDecimal getVersion()
     {
         return version;
     }
 
-    public void setVersion(Integer version)
+    public void setVersion(BigDecimal version)
     {
         this.version = version;
     }
